success

+++ src/controller/bookmark/controller.ts
@@ -4,6 +4,7 @@ import GetBookmark from '../../type/bookmark/getBookmark';
 import BookmarkService from '../../service/bookmark.service';
 import GetBookmarkList from '../../type/bookmark/getBookmarkList';
 import SaveBookmark from '../../type/bookmark/saveBookmark';
+import GetUser from '../../type/user/getUser';
 
 export const getBookmark: RequestHandler = async (req, res, next) => {
   try {
@@ -18,9 +19,10 @@ export const getBookmark: RequestHandler = async (req, res, next) => {
 
 export const getBookmarkList: RequestHandler = async (req, res, next) => {
   try {
-    const id = Number(req.session.id);
+    const id = Number(req.session.user?.id);
     if (!id) throw new BadRequestError('히스토리 목록을 불러올 수 없습니다.');
-    const bookmarkList: GetBookmarkList = await BookmarkService.getBookmarkList(id);
+    const bookmarkList: GetBookmarkList =
+      await BookmarkService.getBookmarkList(id);
     res.status(201).json(bookmarkList);
   } catch (error) {
     next(error);
@@ -29,11 +31,11 @@ export const getBookmarkList: RequestHandler = async (req, res, next) => {
 
 export const saveBookmark: RequestHandler = async (req, res, next) => {
   try {
-    const user = req.session;
+    const user = req.session.user;
     const { placeId } = req.body;
     if (!user || !placeId) throw new BadRequestError('히스토리 저장 실패');
 
-    const createBookmark: SaveBookmark = { user, placeId };
+    const createBookmark: SaveBookmark = { user: user as GetUser, placeId };
     await BookmarkService.saveBookmark(createBookmark);
     res.status(201).send('히스토리가 저장 되었습니다.');
   } catch (error) {
@@ -44,7 +46,8 @@ export const saveBookmark: RequestHandler = async (req, res, next) => {
 export const deleteBookmark: RequestHandler = async (req, res, next) => {
   try {
     const bookmarkId = Number(req.params.id);
-    const userId = Number(req.session.id);
+    const userId = Number(req.session.user?.id);
+    if (!userId || !bookmarkId) throw new BadRequestError('유저 정보가 없음.');
 
     await BookmarkService.deleteBookmark(bookmarkId, userId);
