+++ src/controller/history/controller.ts
@@ -1,10 +1,15 @@
 import { RequestHandler } from 'express';
-import UserService from '../../service/user.service';
 import { BadRequestError } from '../../util/customErrors';
 import GetHistory from '../../type/history/getHistory';
 import HistoryService from '../../service/history.service';
 import GetHistoryList from '../../type/history/getHistoryList';
 import SaveHistory from '../../type/history/saveHistory';
+import GetUser from '../../type/user/getUser';
+declare module 'express-session' {
+  export interface SessionData {
+    user: GetUser;
+  }
+}
 
 export const getHistory: RequestHandler = async (req, res, next) => {
   try {
@@ -19,9 +24,8 @@ export const getHistory: RequestHandler = async (req, res, next) => {
 
 export const getHistoryList: RequestHandler = async (req, res, next) => {
   try {
-    const id = Number(req.session.id);
+    const id = Number(req.session.user?.id);
     if (!id) throw new BadRequestError('히스토리 목록을 불러올 수 없습니다.');
-    const user = await UserService.getUserById(id);
     const historyList: GetHistoryList = await HistoryService.getHistoryList(id);
     res.status(201).json(historyList);
   } catch (error) {
@@ -31,12 +35,12 @@ export const getHistoryList: RequestHandler = async (req, res, next) => {
 
 export const saveHistory: RequestHandler = async (req, res, next) => {
   try {
-    const user = req.session;
+    const user = req.session.user;
     const { placeId } = req.body;
     if (!user || !placeId) throw new BadRequestError('히스토리 저장 실패');
 
-    const createHistory: SaveHistory = { user, placeId };
-    const history = await HistoryService.saveHistory(createHistory);
+    const createHistory: SaveHistory = { user: user as GetUser, placeId };
+    await HistoryService.saveHistory(createHistory);
 
     res.status(201).send('히스토리가 저장 되었습니다.');
   } catch (error) {
@@ -47,7 +51,8 @@ export const saveHistory: RequestHandler = async (req, res, next) => {
 export const deleteHistory: RequestHandler = async (req, res, next) => {
   try {
     const historyId = Number(req.params.id);
-    const userId = Number(req.session.id);
+    const userId = Number(req.session.user?.id);
+    if (!userId || !historyId) throw new BadRequestError('유저 정보가 없음.');
 
     await HistoryService.deleteHistory(historyId, userId);
